Update/footer

+++ smactest-app/src/components/testData.js
@@ -10,15 +10,24 @@ const TestData = (props)=>{
     }
 
     function deleteData(){
-        deleteDocument(props.id);
+        var uInput = prompt('Type: YES to OBLITERATE Heretic');
+        if (uInput.toUpperCase() === 'YES'){
+          alert("Purging Document");
+          deleteDocument(props.id);
+        }
+        else{
+          alert("Cancelling Purge");
+        }
+        
+        
     }
 
     function getPayType(){
         if (props.PayType === 0){
-            return "Cash";
+            return "Cash/ Debit";
         }
         else if (props.PayType === 1){
-            return "Debit";
+            return "E-transfer";
         }
         else if (props.PayType === 2){
             return "Account";
